+++ priv/templates/database/openmaize_ecto_test.exs
@@ -1,8 +1,8 @@
-defmodule <%= base %>.EctoDBTest do
+defmodule <%= base %>.OpenmaizeEctoTest do
   use ExUnit.Case
   use Plug.Test
 
-  alias <%= base %>.{EctoDB, Repo, User}
+  alias <%= base %>.{OpenmaizeEcto, Repo, User}
 
   test "easy password results in an error being added to the changeset" do
     user = %{email: "bill@mail.com", username: "bill", role: "user", password: "easytoguess",
@@ -16,7 +16,7 @@ defmodule <%= base %>.EctoDBTest do
     user = Map.merge(%User{},
                      %{username: "bill", role: "user",
                        confirmation_token: nil, confirmation_sent_at: nil})
-    changeset = EctoDB.add_confirm_token(user, "lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw")
+    changeset = OpenmaizeEcto.add_confirm_token(user, "lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw")
     assert changeset.changes.confirmation_token
     assert changeset.changes.confirmation_sent_at
   end
@@ -25,15 +25,15 @@ defmodule <%= base %>.EctoDBTest do
     user = Map.merge(%User{},
                      %{email: "reg@mail.com", username: "reg", role: "user", password: "h4rd2gU3$$",
                        phone: "081755555", confirmed_at: Ecto.DateTime.utc})
-    changeset = EctoDB.add_reset_token(user, "lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw")
+    changeset = OpenmaizeEcto.add_reset_token(user, "lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw")
     assert changeset.changes.reset_token
     assert changeset.changes.reset_sent_at
   end
 
   test "check time" do
-    assert EctoDB.check_time(Ecto.DateTime.utc, 60)
-    refute EctoDB.check_time(Ecto.DateTime.utc, -60)
-    refute EctoDB.check_time(nil, 60)
+    assert OpenmaizeEcto.check_time(Ecto.DateTime.utc, 60)
+    refute OpenmaizeEcto.check_time(Ecto.DateTime.utc, -60)
+    refute OpenmaizeEcto.check_time(nil, 60)
   end
 
 end